backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minimax-scraper"
version = "0.1.0"
description = "Markdown documentation scraper with OS-like browser UI"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "fastapi[standard]>=0.115.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdownify>=0.14.0",
    "trafilatura>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.21.0",
    "pydantic-settings>=2.7.0",
    "defusedxml>=0.7.0",
    "openai>=1.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"app/api/*.py" = ["B008"]  # Depends() in defaults is standard FastAPI DI pattern

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
